refactor connector to {send,receive} data by specifying only peer id
Currently, tentacle expects session id whereever we want to send
or receive data. This is not ideal in situations like NAT traversal.
In NAT traversal, we normally have a few connections to the remote
peer, e.g. relayed connection, direct connection established by hole
punching, reversed connection when connection from us to them failed
but connection in reverse direction succeeded.
In this case, we can not determine which session id to use easily.
We should instead hide session ids as implementation details and let
the users send data to remote peer by just specifying peer ids.
This is in line with libp2p's behaviour, which has a peer store that
stores all the addresses information of the other peer and in order
to create a new stream to the other peer we only need the peer id.